% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TargetReport.R
\name{CenterQuery}
\alias{CenterQuery}
\title{Perform search and site selection for a single coordinate pair}
\usage{
CenterQuery(
  x,
  key,
  species_user,
  max_sites,
  dist,
  back,
  hotspot,
  include_provisional,
  max_tries,
  timeout_sec,
  verbose,
  drop_patterns
)
}
\arguments{
\item{x}{list with center coordinate information (see Details)}

\item{key}{character vector of user's eBird API key}

\item{species_user}{data frame of species that user has observed}

\item{max_sites}{integer maximum number of sites to return for each pair of
coordinates defined in \code{centers}}

\item{dist}{numeric radius (in kilometers) of area from each point defined
by coordinates in \code{centers} from which to return recent observations}

\item{back}{integer number of days back to search for observations}

\item{hotspot}{logical indicating whether or not to restrict results to
hotspot locations}

\item{include_provisional}{logical indicating whether not to include
observations which have not yet been reviewed}

\item{max_tries}{integer maximum number of query attempts to try (only for
expert use)}

\item{timeout_sec}{integer time to allow before query is aborted (only for
expert use)}

\item{verbose}{logical determining whether or not to print messages during
queries}

\item{drop_patterns}{character vector of patterns in species' names to
exclude certain species from consideration, such as domesticated species,
hybrids, and observations not identified to species level (e.g.
"Toxostoma sp.")}
}
\value{
list of query results.
#' \itemize{
 \item{"center_info"}{List including latitude, longitude, and center name}
 \item{"target_sites"}{List of target sites and potential species}
}
}
\description{
Perform search and site selection for a single coordinate pair
}
\details{
The first argument of this function must be a list of data frames,
each with a single row of data and at least three columns:
\itemize{
 \item{"lat"}{Latitude in decimal degrees}
 \item{"lng"}{Longitude in decimal degrees}
 \item{"name"}{Name of center to use in output report}
}
}
