% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DropPatterns.R
\name{DropPatterns}
\alias{DropPatterns}
\title{Drops any species with given patterns in name}
\usage{
DropPatterns(
  data,
  patterns = c("sp.", "/", "Domestic type", "hybrid"),
  colname = "comName"
)
}
\arguments{
\item{data}{a data.frame with observations, likely the object from a call to
\code{RecentNearby}}

\item{patterns}{character vector of patterns in \code{colname} column to
omit}

\item{colname}{character vector indicating column with names to search for
\code{patterns}}
}
\value{
Copy of \code{data} with any rows that has values in \code{colname}
that match values in \code{patterns}
}
\description{
Drops any species with given patterns in name
}
\details{
This utility function provides a means of omitting observations of
"other taxa" such as domestics, hybrids, and "sp." observations. If
\code{patterns = NULL}, no rows will be dropped.
}
