% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/play-game.R
\name{game_of_life}
\alias{game_of_life}
\title{Play the game of life}
\usage{
game_of_life(matrix, n_iters, speed_limit = 0.2)
}
\arguments{
\item{matrix}{starting matrix}

\item{n_iters}{number of iterations}

\item{speed_limit}{How long to wait between each iteration}
}
\value{
plots the game of life
}
\description{
Play the game of life
}
\examples{

game_matrix <- life_matrix(length = 40, pattern_length = 15)

game_of_life(game_matrix, n_iters = 50, speed_limit = 0.1)

}
