% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/play-game.R
\name{life_matrix}
\alias{life_matrix}
\title{Create a matrix for playing game of life}
\usage{
life_matrix(length = 50, pattern_length = 4)
}
\arguments{
\item{length}{The length of the rows and columns of the entire matrix}

\item{pattern_length}{The length of the rows and columns of the pattern that will populate the matrix}
}
\value{
a matrix made up of only 1s and 0s
}
\description{
Create a matrix for playing game of life
}
\examples{

game_matrix <- life_matrix(length = 60, pattern_length = 15)

}
