% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse.R
\name{parse_cells}
\alias{parse_cells}
\title{Experimental function to parse .cells file from Life Wiki}
\usage{
parse_cells(con)
}
\arguments{
\item{con}{Connection to the file}
}
\value{
a matrix
}
\description{
Experimental function to parse .cells file from Life Wiki
}
\examples{

gosper <- parse_cells("https://conwaylife.com/patterns/gosperglidergun.cells")

}
