% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.recent_obs.R
\name{print.recent_obs}
\alias{print.recent_obs}
\title{Print method for recent_obs object}
\usage{
\method{print}{recent_obs}(obj)
}
\arguments{
\item{obj}{an object of class "recent_obs"}
}
\description{
Print method for recent_obs object
}
