% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RecentNearby.R
\name{RecentNearby}
\alias{RecentNearby}
\title{Retrieve recent nearby observations}
\usage{
RecentNearby(
  key,
  lat = 32.241,
  lng = -110.938,
  dist = 50,
  back = 4,
  hotspot = TRUE,
  include_provisional = FALSE,
  max_tries = 5,
  timeout_sec = 30,
  verbose = TRUE
)
}
\arguments{
\item{key}{character eBird API key}

\item{lat}{numeric latitude; use negative values for southern latitudes
(i.e. -46.86, \emph{not} "46.86 S)}

\item{lng}{numeric longitude; use negative values for western
longitudes (i.e. -72.08, \emph{not} "72.08 W")}

\item{dist}{numeric radius (in kilometers) of area from center point
given by \code{lat} and \code{lng} from which to return recent observations}

\item{back}{integer number of days back to search for observations}

\item{hotspot}{logical indicating whether or not to restrict results to
hotspot locations}

\item{include_provisional}{logical indicating whether not to include
observations which have not yet been reviewed}

\item{max_tries}{integer maximum number of query attempts to try}

\item{timeout_sec}{integer time to allow before query is aborted}

\item{verbose}{logical determining whether or not to print messages during
queries}
}
\value{
An object of class "recent_obs" with the following elements:
\describe{
  \item{query_type}{the type of query performed}
  \item{query_parameters}{list of query parameters passed in request}
  \item{obs}{data frame of observations returned from query; if no
  observations are returned, \code{obs} is NULL}
}
}
\description{
Retrieve recent nearby observations
}
\details{
The function uses the eBird API (see \url{https://documenter.getpostman.com/view/664302/S1ENwy59})
to query recent citings. Queries to the eBird API require a user key; more
information on obtaining a key can be found at the eBird API documentation.
}
\examples{
\dontrun{
  # Read eBird key in from file
  key <- scan(file = "ebird-key.txt", what = "character")
  # Search for observations 5 km from lat/lng coordinates
  recent <- RecentNearby(key = key, lat = 32.28, lng = -111.02, dist = 5)
}

}
