% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.recent_obs.R
\name{summary.recent_obs}
\alias{summary.recent_obs}
\title{Summary of recent_obs object}
\usage{
\method{summary}{recent_obs}(object, ...)
}
\arguments{
\item{object}{An object of class "recent_obs".}
}
\value{
No return value, called to print summary of a \code{recent_obs} 
object
}
\description{
Summary of recent_obs object
}
\keyword{internal}
