% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TargetReport.R
\name{TargetReport}
\alias{TargetReport}
\title{Create report for sites to target}
\usage{
TargetReport(
  centers,
  key_file,
  list_file,
  center_names = NULL,
  max_sites = 5,
  dist = 50,
  back = 4,
  hotspot = TRUE,
  include_provisional = FALSE,
  max_tries = 5,
  timeout_sec = 30,
  verbose = TRUE,
  drop_patterns = c("sp.", "/", "Domestic type", "hybrid")
)
}
\arguments{
\item{centers}{numeric vector or matrix of latitude and longitude
coordinates; vector should be of length 2, while matrix should have two
columns}

\item{key_file}{path to file with eBird API key}

\item{list_file}{path to file of user's observations; this should be a csv
file of observations downloaded from your My eBird page at
\url{https://ebird.org/myebird}}

\item{center_names}{optional character vector of names to use for each pair
of latitude and longitude coordinates in \code{centers}}

\item{max_sites}{integer maximum number of sites to return for each pair of
coordinates defined in \code{centers}}

\item{dist}{numeric radius (in kilometers) of area from each point defined
by coordinates in \code{centers} from which to return recent observations}

\item{back}{integer number of days back to search for observations}

\item{hotspot}{logical indicating whether or not to restrict results to
hotspot locations}

\item{include_provisional}{logical indicating whether not to include
observations which have not yet been reviewed}

\item{max_tries}{integer maximum number of query attempts to try (only for
expert use)}

\item{timeout_sec}{integer time to allow before query is aborted (only for
expert use)}

\item{verbose}{logical determining whether or not to print messages during
queries}

\item{drop_patterns}{character vector of patterns in species' names to
exclude certain species from consideration, such as domesticated species,
hybrids, and observations not identified to species level (e.g.
"Toxostoma sp.")}
}
\description{
Create report for sites to target
}
\details{
The function uses the eBird API (see \url{https://documenter.getpostman.com/view/664302/S1ENwy59})
to build the target report. Queries to the eBird API require a user key; more
information on obtaining a key can be found at the eBird API documentation.
}
