% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MapSites.R
\name{MapSites}
\alias{MapSites}
\title{Create map of sites identified with highest number of missing species}
\usage{
MapSites(sites, center_lng = NULL, center_lat = NULL)
}
\arguments{
\item{sites}{data frame with at least the following columns:
\describe{
  \item{locName}{character name of site}
  \item{locId}{eBird identifier for site}
  \item{num_new}{integer number of new species at site}
  \item{lat}{double latitude in decimal degrees}
  \item{lng}{double longitude in decimal degrees}
}}

\item{center_lat}{numeric decimal degrees latitude of}
}
\value{
a ggmap object
}
\description{
Create map of sites identified with highest number of missing species
}
\details{
The function is primarily used by \code{SitesReport} via the 
template RMarkdown file used to build reports. It is not intended for 
standalone use.
}
\examples{
\dontrun{
  # Create data frame with required columns
  localities <- data.frame(locName = c("Sweetwater Wetlands", "Reid Park"), 
  locID = c("L208918", "L227274"), num_new = c(3, 5), 
  lat = c(32.279, 32.210), lng = c(-111.022, -110.924))
  lifeR:::MapSites(sites = localities)
}
}
\keyword{internal}
