% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.recent_obs.R
\name{print.recent_obs}
\alias{print.recent_obs}
\title{Print recent_obs object}
\usage{
\method{print}{recent_obs}(x, ...)
}
\arguments{
\item{x}{An object of class \code{recent_obs}.}
}
\description{
Print recent_obs object
}
\keyword{internal}
