% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SitesReport.R
\name{SitesReport}
\alias{SitesReport}
\title{Create report for sites with most unseen species}
\usage{
SitesReport(
  centers,
  ebird_key,
  species_seen,
  center_names = NULL,
  report_filename = "Goals-Report",
  report_dir = getwd(),
  report_format = c("html", "pdf"),
  max_sites = 5,
  dist = 50,
  back = 4,
  hotspot = TRUE,
  include_provisional = FALSE,
  max_tries = 5,
  timeout_sec = 30,
  verbose = FALSE,
  drop_patterns = c("sp.", "/", "Domestic type", "hybrid")
)
}
\arguments{
\item{centers}{numeric vector or matrix of latitude and longitude
coordinates; vector should be of length 2, e.g. 
\code{c(latitude, longitude)}, while matrix should have two columns (first 
column is latitude, second column is longitude)}

\item{ebird_key}{character vector with eBird API key}

\item{species_seen}{character vector of species that have already been seen}

\item{center_names}{optional character vector of names to use for each pair
of latitude and longitude coordinates in \code{centers}}

\item{report_filename}{name of output file without file extension (see 
\code{report_format})}

\item{report_dir}{destination folder for the output file}

\item{report_format}{file format for report; takes one of two values: "html" 
and "pdf"}

\item{max_sites}{integer maximum number of sites to return for each pair of
coordinates defined in \code{centers}}

\item{dist}{numeric radius (in kilometers) of area from each point defined
by coordinates in \code{centers} from which to return recent observations}

\item{back}{integer number of days back to search for observations}

\item{hotspot}{logical indicating whether or not to restrict results to
hotspot locations}

\item{include_provisional}{logical indicating whether not to include
observations which have not yet been reviewed}

\item{max_tries}{integer maximum number of query attempts to try (only for
expert use)}

\item{timeout_sec}{integer time to allow before query is aborted (only for
expert use)}

\item{verbose}{logical determining whether or not to print messages during
queries}

\item{drop_patterns}{character vector of patterns in species' names to
exclude certain species from consideration, such as domesticated species,
hybrids, and observations not identified to species level (e.g.
"Toxostoma sp.")}
}
\value{
Silently returns the results of queries as a list. Each element of 
the list is a two element list with named elements:
\describe{
  \item{results_list}{a list where each element is a list of the results of 
  queries for a center}
  \item{report_details}{a list containing the settings used to build this 
  report, such as days back and distances}
}
}
\description{
Create report for sites with most unseen species
}
\details{
The function uses the eBird API (see \url{https://documenter.getpostman.com/view/664302/S1ENwy59})
to build the report. Queries to the eBird API require a user key; more
information on obtaining a key can be found at the eBird API documentation.
}
